op-acceptance-tests: rename halt package to reorg, consolidate tests
- Rename halt/ package to reorg/ (reflects actual behavior)
- Move invalid_message_replacement_test.go into reorg package
- Delete obsolete invalid_message_halt_test.go (superseded by reorg test)

The halt test tested old behavior where invalid messages caused the chain
to halt. With block invalidation, the chain now rewinds and replaces
the invalid block with a deposits-only block.
