+++ op-acceptance-tests/tests/supernode/interop/reorg/init_test.go
@@ -1,4 +1,4 @@
-package halt
+package reorg
 
 import (
 	"os"
@@ -8,7 +8,7 @@ import (
 )
 
 // TestMain creates an isolated two-L2 setup with a shared supernode that has interop enabled.
-// This package tests invalid message scenarios that would pollute other tests if run on a shared devnet.
+// This package tests block invalidation and reorg scenarios that would pollute other tests if run on a shared devnet.
 func TestMain(m *testing.M) {
 	_ = os.Setenv("DEVSTACK_L2CL_KIND", "supernode")
 	presets.DoMain(m, presets.WithTwoL2SupernodeInterop(0))
